server/requirements.txt
scipy
numpy
soundfile
pretty_midi
pydub

fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
requests
jinja2
bcrypt 
fastapi-login
httpx
python-jose
itsdangerous